requirements.txt
fastmcp==2.12.5
fastapi==0.119.0
uvicorn==0.38.0
